packages/ingest/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regulasense-ingest"
version = "0.1.0"
description = "Data ingestion module for RegulaSense"
authors = [
    {name = "RegulaSense Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "fredapi>=0.5.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "openai>=1.25.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "tqdm>=4.66.0",
]

[project.scripts]
ingest = "regulasense_ingest.cli:main"

[tool.setuptools]
packages = ["regulasense_ingest", "regulasense_ingest.sources", "regulasense_ingest.utils"]

[tool.setuptools.package-data]
regulasense_ingest = ["py.typed"] 